Add comprehensive tests for PayFlow API and services
- Implement endpoint tests for health check, merchant registration, payment creation, and rate limiting in `endpoints_test.rs`.
- Create SQL fixtures for test data in `test_data.sql`, including merchants, wallets, payments, and balances.
- Develop full integration tests covering health checks, metrics, merchant creation, and payment listing in `full_integration_test.rs`.
- Add payment service tests for creating payments, fee calculations, and payment retrieval in `payment_test.rs`.
- Implement service integration tests for merchant registration, balance operations, and deposit address generation in `services_test.rs`.
- Create standalone unit tests for encryption, key generation, and validation in `standalone_tests.rs`.
- Add utility tests for encryption roundtrip and keypair generation in `utils_test.rs`.
- Implement withdrawal service tests for creating withdrawals, checking minimum amounts, and listing withdrawals in `withdrawal_test.rs`.
- Develop end-to-end workflow tests for payment and withdrawal processes in `workflows_test.rs`.

+++ migrations/20240101000004_balance_management.sql
@@ -1,7 +1,7 @@
 -- Create merchant_balances table
 CREATE TABLE merchant_balances (
     id SERIAL PRIMARY KEY,
-    merchant_id INTEGER NOT NULL REFERENCES merchants(id),
+    merchant_id BIGINT NOT NULL REFERENCES merchants(id),
     crypto_type VARCHAR(50) NOT NULL,
     available_balance DECIMAL(36, 18) NOT NULL DEFAULT 0,
     reserved_balance DECIMAL(36, 18) NOT NULL DEFAULT 0,
@@ -13,7 +13,7 @@ CREATE TABLE merchant_balances (
 -- Create balance_history table
 CREATE TABLE balance_history (
     id SERIAL PRIMARY KEY,
-    merchant_id INTEGER NOT NULL REFERENCES merchants(id),
+    merchant_id BIGINT NOT NULL REFERENCES merchants(id),
     crypto_type VARCHAR(50) NOT NULL,
     amount DECIMAL(36, 18) NOT NULL,
     balance_type VARCHAR(20) NOT NULL, -- 'AVAILABLE' or 'RESERVED'
